requirements-core.txt
# Core scientific stack (pinned, rarely change)
numpy==2.3.5
pandas==3.0.0
scipy==1.17.0
scikit-learn==1.8.0
biopython==1.86
tqdm==4.67.3

# Visualization
matplotlib==3.10.8
seaborn==0.13.2
plotly==6.5.2
kaleido==1.2.0

# ML / analysis
umap-learn==0.5.11
prince==0.16.5
kneebow==1.0.4
wordcloud==1.9.6

# Web scraping
selenium==4.40.0
webdriver-manager==4.0.2
beautifulsoup4==4.14.3

# Config / utilities
pyyaml==6.0.3
